Refactor config and AWS Client functions
These functions work better as their own modules, this lets them
be re-used in other parts of the function, such as the Kinesis
writer. Doing so also fixes a glitch with loading aws credentials
from the config file

Several tests had to be updated to reflect this new structure
